refactor: User 도메인 파일 위치 변경
