+++ src/main/java/com/blog/domain/auth/login/standard/service/LoginService.java
@@ -5,8 +5,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import com.blog.domain.auth.jwt.service.TokenProvider;
-import com.blog.domain.user.domain.User;
-import com.blog.domain.user.repository.UserRepository;
+import com.blog.domain.user.domain.entity.User;
+import com.blog.domain.user.domain.repository.UserRepository;
 import com.blog.domain.auth.jwt.encoder.PasswordEncoder;
 import com.blog.domain.auth.login.standard.dto.AuthenticationResultDTO;
 import com.blog.domain.auth.login.standard.dto.LoginRequestDTO;
