+++ src/main/java/com/blog/domain/user/application/mapper/UserRowMapper.java
@@ -1,11 +1,11 @@
-package com.blog.domain.user.mapper;
+package com.blog.domain.user.application.mapper;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
 import org.springframework.jdbc.core.RowMapper;
 
-import com.blog.domain.user.domain.User;
+import com.blog.domain.user.domain.entity.User;
 
 public class UserRowMapper implements RowMapper<User> {
 	@Override
